refactor: IntegrationFixture 패키지 변경
